refactor(lns-cli): move volume args (VolumeArgs/VolumeCommand/…) into the volume module
