+++ crates/lns-cli/src/volume/mod.rs
@@ -3,7 +3,6 @@ use std::io::{BufRead, Write};
 use anyhow::{Result, bail};
 use lns_ipc::{Request, Response, VolumeInfo};
 
-use crate::cli::VolumeCommand;
 use crate::command::{CommandSpec, subcommand};
 use crate::integration::LocalBoxFuture;
 
@@ -11,9 +10,54 @@ mod real;
 
 pub use real::RealVolumeService;
 
+#[derive(clap::Args)]
+pub struct VolumeArgs {
+    #[command(subcommand)]
+    pub command: VolumeCommand,
+}
+
+#[derive(clap::Subcommand)]
+pub enum VolumeCommand {
+    #[command(about = "List named volumes with their on-disk size, age, and holder.")]
+    Ls,
+    #[command(about = "Create a named volume ahead of its first `lns run -v` attach.")]
+    Create(VolumeNameArg),
+    #[command(about = "Show a volume's details as JSON.")]
+    Inspect(VolumeNameArg),
+    #[command(about = "Remove a named volume; refused while a run holds it.")]
+    Rm(VolumeNameArg),
+    #[command(about = "Remove every volume not attached to a running sandbox.")]
+    Prune(VolumePruneArgs),
+}
+
+#[derive(clap::Args)]
+pub struct VolumeNameArg {
+    #[arg(
+        value_parser = parse_volume_name,
+        help = "Volume name, as used with `lns run -v name:/path`."
+    )]
+    pub name: String,
+}
+
+#[derive(clap::Args)]
+pub struct VolumePruneArgs {
+    #[arg(
+        short = 'f',
+        long,
+        default_value_t = false,
+        help = "Skip the confirmation prompt."
+    )]
+    pub force: bool,
+}
+
+fn parse_volume_name(s: &str) -> Result<String, String> {
+    lns_ipc::validate_volume_name(s)?;
+    Ok(s.to_string())
+}
+
 pub fn augment(app: clap::Command) -> clap::Command {
     app.subcommand(
-        subcommand::<crate::cli::VolumeArgs>("volume")
+        subcommand::<VolumeArgs>("volume")
             .about("Manage the named volumes used with `lns run -v` (`docker volume`-style)."),
     )
 }
@@ -263,8 +307,8 @@ mod tests {
         Ok((code, String::from_utf8(buf).unwrap()))
     }
 
-    fn name_arg(name: &str) -> crate::cli::VolumeNameArg {
-        crate::cli::VolumeNameArg {
+    fn name_arg(name: &str) -> VolumeNameArg {
+        VolumeNameArg {
             name: name.to_string(),
         }
     }
@@ -276,7 +320,7 @@ mod tests {
             VolumeCommand::Create(name_arg("v")),
             VolumeCommand::Inspect(name_arg("v")),
             VolumeCommand::Rm(name_arg("v")),
-            VolumeCommand::Prune(crate::cli::VolumePruneArgs { force: true }),
+            VolumeCommand::Prune(VolumePruneArgs { force: true }),
         ] {
             let svc = CannedService::with([Some(Response::Pong)]);
             let err = run_cmd(&cmd, &svc).await.unwrap_err().to_string();
@@ -288,7 +332,7 @@ mod tests {
     async fn prune_without_force_reads_eof_as_decline() {
         let svc = CannedService::with([]);
         let (code, out) = run_cmd(
-            &VolumeCommand::Prune(crate::cli::VolumePruneArgs { force: false }),
+            &VolumeCommand::Prune(VolumePruneArgs { force: false }),
             &svc,
         )
         .await
