+++ crates/lns-cli/src/volume/real.rs
@@ -9,7 +9,7 @@ use crate::integration::LocalBoxFuture;
 
 pub fn run<'a>(matches: &'a clap::ArgMatches, ctx: RunCtx<'a>) -> RunFuture<'a> {
     Box::pin(async move {
-        let args = crate::cli::VolumeArgs::from_arg_matches(matches)?;
+        let args = super::VolumeArgs::from_arg_matches(matches)?;
         crate::service::require_running().await;
         let svc = RealVolumeService::new(crate::service::socket_path()?);
         let mut out = ctx.out;
